:art: Refactor URI processing and enhance event bus integration for improved handling of SiYuan URIs
