+++ app/src/constants.ts
@@ -31,6 +31,9 @@ export abstract class Constants {
     public static readonly SIYUAN_DROP_TAB: string = "application/siyuan-tab";
     public static readonly SIYUAN_DROP_EDITOR: string = "application/siyuan-editor";
 
+    // App 内部全局事件
+    public static readonly SIYUAN_APP_EVENT_OPEN_URI: string = "open-uri";
+
     // 渲染进程调主进程
     public static readonly SIYUAN_CMD: string = "siyuan-cmd";
     public static readonly SIYUAN_GET: string = "siyuan-get";
